Expose the geomapper interface (ecoinvent locations - REMIND regions) to be used by other packages or interanlly.
